requirements.txt
django==1.11.9
djangorestframework
django-cors-headers
django-filter